Move peer-link client mDNS wake listener into _wake.py
